+++ doc/CHANGES.rst
@@ -19,6 +19,7 @@ Features Added
 - Updated distributions:
 
   - Acquisition = 2.13.7
+  - Products.BTreeFolder2 = 2.13.2
   - Products.ZCatalog = 2.13.7
 
 2.13.5 (2011-02-23)